requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

lightgbm==4.0.0
xgboost==1.7.6
catboost==1.2.0

torch==2.0.1
torchvision==0.15.2

Pillow==10.0.0
opencv-python==4.8.0.74

tqdm==4.65.0
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
optuna==3.3.0


joblib==1.3.2